feat(schema): background index builds with DEFINE INDEX ... CONCURRENTLY
IndexDefinition::with_concurrently emits the directive; info_for_index_surql
plus IndexBuildStatus read the build progress back.

v3.0.5 accepts CONCURRENTLY and echoes the index without it, so the parser
reports concurrently: false and nothing compares the member -- otherwise
every background-built index would diff forever.

IndexDefinition and its enums and builders moved from schema::table to a
new schema::index (re-exported) so both files stay under 1000 lines.
